move call_repeatedly into utils
